refactor(ontologies): extract searchWellKnownOntologies, eliminate filter duplication
Single exported function in wellKnownOntologies.ts. AutoComplete and
searchOntologies MCP tool both call it — one code path.
